路由 restful改造

+++ pet-project/handler/record_handler.go
@@ -124,7 +124,9 @@ func DeletePetInfo(c *gin.Context) {
 // GetCommonCategories 获取宠物分类
 func GetCommonCategories(c *gin.Context) {
 	var petActionList []models.RecordCategory
-	result := db.DB.Model(&models.RecordCategory{}).Find(&petActionList)
+	result := db.DB.Model(&models.RecordCategory{}).
+		Where("user_id IS NULL").
+		Find(&petActionList)
 	if result.Error != nil {
 		response.Fail(c, response.ApiCode.QueryErr, response.ApiMsg.QueryErr)
 		return
@@ -133,14 +135,44 @@ func GetCommonCategories(c *gin.Context) {
 }
 
 func CreateCommonCategory(c *gin.Context) {
-	var recordCategory models.RecordCategory
-	if err := c.ShouldBind(&recordCategory); err != nil {
+	userId := c.MustGet("userId").(uint)
+	if userId == 1 {
+		var recordCategory models.RecordCategory
+		if err := c.ShouldBind(&recordCategory); err != nil {
+			response.Fail(c, response.ApiCode.ParamErr, response.ApiMsg.ParamErr)
+			return
+		}
+		recordCategory.UserId = nil
+		result := db.DB.Create(&recordCategory)
+		if result.Error != nil {
+			response.Fail(c, response.ApiCode.CreateErr, response.ApiMsg.CreateErr)
+			return
+		}
+		response.Success(c, nil)
+	} else {
+		response.Fail(c, response.ApiCode.RejectErr, response.ApiMsg.RejectErr)
+	}
+}
+
+// CreateCommonCategoryList 批量添加宠物行为
+func CreateCommonCategoryList(c *gin.Context) {
+	var categories []models.RecordCategory
+	if err := c.ShouldBind(&categories); err != nil {
 		response.Fail(c, response.ApiCode.ParamErr, response.ApiMsg.ParamErr)
 		return
 	}
-
-	result := db.DB.Create(&recordCategory)
-	if result.Error != nil {
+	var values []bool
+	for i := range categories {
+		var item = categories[i]
+		item.UserId = nil
+		result := db.DB.Omit(clause.Associations).Create(&categories[i])
+		if result.Error != nil {
+			values = append(values, false)
+		} else {
+			values = append(values, true)
+		}
+	}
+	if slices.Contains(values, false) {
 		response.Fail(c, response.ApiCode.CreateErr, response.ApiMsg.CreateErr)
 		return
 	}
@@ -149,21 +181,35 @@ func CreateCommonCategory(c *gin.Context) {
 
 // DeleteCommonCategory 删除宠物行为
 func DeleteCommonCategory(c *gin.Context) {
-	id := c.Param("id")
-	result := db.DB.Delete(&models.RecordCategory{}, "id = ?", id)
-	if result.Error != nil {
-		response.Fail(c, response.ApiCode.QueryErr, response.ApiMsg.QueryErr)
-		return
+	userId := c.MustGet("userId").(uint)
+	if userId == 1 {
+		id := c.Param("id")
+		result := db.DB.Delete(&models.RecordCategory{}, "id = ? AND user_id IS NULL", id)
+		if result.Error != nil {
+			response.Fail(c, response.ApiCode.QueryErr, response.ApiMsg.QueryErr)
+			return
+		}
+		response.Success(c, nil)
+	} else {
+		response.Fail(c, response.ApiCode.RejectErr, response.ApiMsg.RejectErr)
 	}
-	response.Success(c, nil)
+
 }
 
 // GetRecordCategoryList 获取宠物行为列表
 func GetRecordCategoryList(c *gin.Context) {
-	var uerId = c.MustGet("userId").(uint)
+	var userId = c.MustGet("userId").(uint)
+	var pageModel models.PageModel
+	if err := c.ShouldBindQuery(&pageModel); err != nil {
+		response.Fail(c, response.ApiCode.ParamErr, response.ApiMsg.ParamErr)
+		return
+	}
+	offset := (pageModel.PageNum - 1) * pageModel.PageSize
 	var petActionList []models.RecordCategory
 	result := db.DB.Model(&models.RecordCategory{}).
-		Where("user_id = ?", uerId).
+		Where("user_id IS NULL").Or("user_id = ?", userId).
+		Offset(offset).
+		Limit(pageModel.PageSize).
 		Find(&petActionList)
 	if result.Error != nil {
 		response.Fail(c, response.ApiCode.QueryErr, response.ApiMsg.QueryErr)
@@ -174,12 +220,13 @@ func GetRecordCategoryList(c *gin.Context) {
 
 // CreateRecordCategory 添加宠物行为
 func CreateRecordCategory(c *gin.Context) {
+	userId := c.MustGet("userId").(uint)
 	var recordCategory models.RecordCategory
 	if err := c.ShouldBind(&recordCategory); err != nil {
 		response.Fail(c, response.ApiCode.ParamErr, response.ApiMsg.ParamErr)
 		return
 	}
-
+	recordCategory.UserId = &userId
 	result := db.DB.Create(&recordCategory)
 	if result.Error != nil {
 		response.Fail(c, response.ApiCode.CreateErr, response.ApiMsg.CreateErr)
@@ -189,33 +236,11 @@ func CreateRecordCategory(c *gin.Context) {
 }
 
 func DeleteRecordCategory(c *gin.Context) {
+	userId := c.MustGet("userId").(uint)
 	id := c.Param("id")
-	result := db.DB.Delete(&models.RecordCategory{}, "id = ?", id)
+	result := db.DB.Delete(&models.RecordCategory{}, "id = ? AND user_id = ?", id, userId)
 	if result.Error != nil {
 		response.Fail(c, response.ApiCode.QueryErr, response.ApiMsg.QueryErr)
-		return
-	}
-	response.Success(c, nil)
-}
-
-// CreateCategoryList 批量添加宠物行为
-func CreateCategoryList(c *gin.Context) {
-	var recordList []models.RecordCategory
-	if err := c.ShouldBind(&recordList); err != nil {
-		response.Fail(c, response.ApiCode.ParamErr, response.ApiMsg.ParamErr)
-		return
-	}
-	var values []bool
-	for i := range recordList {
-		result := db.DB.Omit(clause.Associations).Create(&recordList[i])
-		if result.Error != nil {
-			values = append(values, false)
-		} else {
-			values = append(values, true)
-		}
-	}
-	if slices.Contains(values, false) {
-		response.Fail(c, response.ApiCode.CreateErr, response.ApiMsg.CreateErr)
 		return
 	}
 	response.Success(c, nil)
